Make dsp code work with updated starch
